feat(client): Setup project filtering + re-organized projects section
