+++ src/app/sections/projects/index.tsx
@@ -5,6 +5,10 @@ export const revalidate = 60;
 
 export default async function Projects() {
   const projects = await api.projects.list({ onlyPublished: true });
+  const filterProperties = await api.projects.filterProperties({
+    onlyPublished: true,
+    projects,
+  });
 
   return (
     <section className="min-h-screen">
@@ -14,7 +18,12 @@ export default async function Projects() {
       <p className="text-neutral-500 text-center text-xl mb-4">
         Check out what I&apos;ve been working on!
       </p>
-      <ProjectCards projects={projects} />
+      <ProjectCards
+        projects={projects}
+        allTypes={filterProperties.types}
+        allLanguages={filterProperties.languages}
+        allFrameworks={filterProperties.frameworks}
+      />
     </section>
   );
 }
